requirements.txt
numpy==2.3.0
Pillow==10.1.0
Pillow==11.2.1
streamlit==1.45.1
tensorflow==2.19.0
